restructure stubs to build docs and satisfy lsp
