+++ docs/source/autoapi/index.rst
@@ -7,5 +7,7 @@ This page contains auto-generated API reference documentation [#f1]_.
    :titlesonly:
 
    /autoapi/rpaudio/index
+   /autoapi/effects/index
+   /autoapi/exceptions/index
 
 .. [#f1] Created with `sphinx-autoapi <https://github.com/readthedocs/sphinx-autoapi>`_
