--- docs/source/autoapi/rpaudio/effects/index.rst
@@ -1,17 +1,0 @@
-rpaudio.effects
-===============
-
-.. py:module:: rpaudio.effects
-
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 1
-
-   /autoapi/rpaudio/effects/ChangeSpeed/index
-   /autoapi/rpaudio/effects/FadeIn/index
-   /autoapi/rpaudio/effects/FadeOut/index
-
-
