--- docs/source/autoapi/rpaudio/rpaudio_exceptions/index.rst
@@ -1,15 +1,0 @@
-rpaudio.rpaudio_exceptions
-==========================
-
-.. py:module:: rpaudio.rpaudio_exceptions
-
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 1
-
-   /autoapi/rpaudio/rpaudio_exceptions/EffectConflictException/index
-
-
