+++ examples/AudioChannel/single_channel_effects.py
@@ -1,7 +1,6 @@
-import python.rpaudio.rpaudio as rpaudio
-from python.rpaudio.rpaudio import FadeIn, FadeOut
+import rpaudio
+from rpaudio.effects import FadeIn, FadeOut
 import asyncio
-from datetime import datetime, timedelta
 
 
 """
