+++ examples/AudioSink/apply_effects.py
@@ -1,6 +1,6 @@
-import python.rpaudio.rpaudio as rpaudio
+import  rpaudio
 import asyncio
-from python.rpaudio.rpaudio import FadeIn, FadeOut, ChangeSpeed
+from rpaudio.effects import FadeIn, FadeOut, ChangeSpeed
 
 kill_audio = False
 AUDIO_FILE = r"C:\Users\16145\Desktop\exc.mp3"
@@ -13,7 +13,6 @@ def on_audio_stop():
 async def play_audio():
     handler = rpaudio.AudioSink(callback=on_audio_stop).load_audio(AUDIO_FILE)
     print(handler.metadata)
-    # handler.try_seek(148.0)
     await asyncio.sleep(0.3)
     handler.set_volume(0.0)
 
@@ -21,7 +20,7 @@ async def play_audio():
     fade_out_effect = FadeOut(duration=10.0, apply_after=10.0)
     speed_up = ChangeSpeed(apply_after=5.0, end_val=1.5, duration=3.0)
 
-    effects_list = [fade_in_effect]
+    effects_list = [fade_in_effect, fade_out_effect, speed_up]
     handler.apply_effects(effects_list)
     handler.play()
